finished blog

+++ posts/views.py
@@ -1,16 +1,25 @@
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.contrib import messages
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
+from django.db.models import Q
+from django.contrib.contenttypes.models import ContentType
 from django.shortcuts import render, get_object_or_404, redirect
-
+from django.utils import timezone
 # Create your views here.
+from comments.forms import CommentForm
+from comments.models import Comment
 from .forms import PostForm
 from .models import Post
 #cоздание поста
 def post_create(request):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    if not request.user.is_authenticated():
+        raise Http404
     form = PostForm(request.POST or None, request.FILES or None)
     if form.is_valid():
         instanse = form.save(commit=False)
+        instanse.user = request.user
         instanse.save()
         messages.success(request, "Успешно создан")
         return HttpResponseRedirect(instanse.get_absolut_url())
@@ -21,16 +30,54 @@ def post_create(request):
     return render(request, "post_form.html", context)
 
 #просмотр поста
-def post_detail(request, id=None):
-    instance = get_object_or_404(Post, id=id)
+def post_detail(request, slug=None):
+    instance = get_object_or_404(Post, slug=slug)
+    if instance.publish > timezone.now().date() or instance.draft:
+        if not request.user.is_staff or not request.user.is_superuser:
+            raise Http404
+    #Связка по ключу
+    
+    initial_data = {
+        "content_type": instance.get_content_type,
+        "object_id": instance.id
+    }
+    form = CommentForm(request.POST or None, initial=initial_data)
+    if form.is_valid():
+        c_type = form.cleaned_data.get("content_type")
+        content_type = ContentType.objects.get(model=c_type)
+        obj_id = form.cleaned_data.get("object_id")
+        content_data = form.cleaned_data.get("content")
+        new_comment, created = Comment.objects.get_or_create(
+                                  user = request.user,
+                                  content_type=content_type,
+                                  object_id=obj_id,
+                                  content=content_data
+                            )
+
+    comments = Comment.objects.filter_by_instance(instance)
+    #########################
+    #Post.object.get(id=instance.id)
     context = {       
-       "instance": instance  
+       "instance": instance,
+       "comments": comments,
+       "comment_form": form  
     }
     return render(request, "post_detail.html", context)
 
 #список постов
 def post_list(request):
-    queryset_list = Post.objects.all()
+    queryset_list = Post.objects.active()
+    if request.user.is_staff or request.user.is_superuser:
+        queryset_list = Post.objects.all()
+
+    query = request.GET.get("q")
+    if query:
+      queryset_list = queryset_list.filter(
+              Q(title__icontains=query) |
+              Q(content__icontains=query) |
+              Q(user__first_name__icontains=query) |
+              Q(user__last_name__icontains=query)
+              ).distinct()
 
     paginator = Paginator(queryset_list, 3) # Show 25 contacts per page
     page_request_var = "page"
@@ -57,7 +104,6 @@ def post_list(request):
 
     context = {
         "object_list": queryset,
-        "title": "List",
         "page_request_var": page_request_var,
         "page_range": page_range
     }
@@ -65,6 +111,10 @@ def post_list(request):
 
 #обновление поста
 def post_update(request, id=None):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    if not request.user.is_authenticated():
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     form = PostForm(request.POST or None, request.FILES or None, instance=instance)
     if form.is_valid():
@@ -80,6 +130,8 @@ def post_update(request, id=None):
 
 #удаление поста
 def post_delete(request, id=None):
+    if not request.user.is_staff or not request.user.superuser:
+        raise Http404
     instance = get_object_or_404(Post, id=id)
     instance.delete()
     messages.success(request, "Успешно Удален")
